update ContrastManager, update gui_hcs.py
