+++ software/configurations/objectives.csv
@@ -1,0 +1,11 @@
+name,magnification,NA,tube_lens_f_mm
+2x,2,0.10,180
+4x,4,0.13,180
+10x,10,0.3,180
+10x (Mitutoyo),10,0.25,200
+20x,20,0.8,180
+20x (Nikon),20,0.45,200
+20x (Boli),20,0.4,180
+40x,40,0.95,180
+50x,50,0.8,180
+60x,60,1.2,180
