+++ software/configurations/sample_formats.csv
@@ -1,0 +1,7 @@
+format,a1_x_mm,a1_y_mm,a1_x_pixel,a1_y_pixel,well_size_mm,well_spacing_mm,number_of_skip,rows,cols
+6,24.55,23.01,290,272,34.94,39.2,0,2,3
+12,24.75,16.86,293,198,22.05,26.0,0,3,4
+24,24.45,22.07,233,210,15.54,19.3,0,4,6
+96,11.31,10.75,171,135,6.21,9.0,0,8,12
+384,12.05,9.05,143,106,3.3,4.5,1,16,24
+1536,11.01,7.87,130,93,1.53,2.25,0,32,48
